# Conflicts:
#	common/include/mm/PageManager.h
#	common/source/kernel/Loader.cpp
#	common/source/mm/PageManager.cpp

+++ common/include/kernel/UserProcess.h
@@ -92,9 +92,6 @@ class UserProcess
 
         ustl::string str() const;
 
-        //UserProcess* current_process_;
-        //UserProcess* getCurrentProcess();
-
-        int waitProcess(size_t pid, int* status, int options);
+        long int waitProcess(long int pid, int* status, int options);
 };
