+++ userspace/libc/src/wait.c
@@ -8,7 +8,7 @@
  */
 pid_t waitpid(pid_t pid, int *status, int options)
 {
-  return __syscall(sc_wait_pid, (size_t )pid, (size_t )status, (size_t )options, 0x0, 0x0);
+  return __syscall(sc_wait_pid, (long int)pid, (size_t )status, (size_t )options, 0x0, 0x0);
 
 }
